refactor: s/$staffSidebar/$sidebar/
